% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_gbf.R
\name{compare_gbf}
\alias{compare_gbf}
\title{Compare current best solution with best neighbour}
\usage{
compare_gbf(y_gbf, neighbour_gbf)
}
\arguments{
\item{y_gbf}{Current best solution}

\item{neighbour_gbf}{Best neighbouring solution}
}
\value{
Updated best solution
}
\description{
Compare current best solution with best neighbour
}
\keyword{internal}
