% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbf_set.R
\name{gbf_set}
\alias{gbf_set}
\title{Compute the generalised Bayes factor score for each solution in the solution set (sol_set)}
\usage{
gbf_set(sol_set, evidence_set, evidence_states, bn_grain)
}
\arguments{
\item{sol_set}{Solution set}

\item{evidence_set}{Set of observed variable names}

\item{evidence_states}{Set of observed variable states}

\item{bn_grain}{Bayesian network as grain object}
}
\value{
Generalised Bayes factor scores for all variables in the solution set (sol_set)
}
\description{
Compute the generalised Bayes factor score for each solution in the solution set (sol_set)
}
\keyword{internal}
