% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_gbf.R
\name{init_gbf}
\alias{init_gbf}
\title{Compute the initialisation values for each target in the set}
\usage{
init_gbf(target_set, evidence_set, evidence_states, bn_grain)
}
\arguments{
\item{target_set}{Set of variables to explore to explain the observed evidence}

\item{evidence_set}{Set of observed variable names}

\item{evidence_states}{Set of observed variable states}

\item{bn_grain}{Bayesian network as grain object}
}
\value{
Starting solutions for each target variable in target_set
}
\description{
Compute the initialisation values for each target in the set
}
