% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_exp.R
\name{minimal_exp}
\alias{minimal_exp}
\title{Return minimal set based on dominance relations}
\usage{
minimal_exp(best_set)
}
\arguments{
\item{best_set}{Set of most relevant explanations}
}
\value{
Minimal explanations
}
\description{
Return minimal set based on dominance relations
}
\keyword{internal}
