% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mre_brute.R
\name{mre_brute}
\alias{mre_brute}
\title{Brute force search to obtain all explanations according to generalised Bayes factor}
\usage{
mre_brute(target_set, evidence_set, evidence_states, bn_grain)
}
\arguments{
\item{target_set}{Set of variables to explore to explain the observed evidence}

\item{evidence_set}{Set of observed variable names}

\item{evidence_states}{Set of observed variable states}

\item{bn_grain}{Bayesian network as grain object}
}
\value{
Dataframe containing all explanations
}
\description{
Brute force search to obtain all explanations according to generalised Bayes factor
}
