% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mre_fwd.R
\name{mre_fwd}
\alias{mre_fwd}
\title{Solve the most relevant explanation in Bayesian networks using a forward search algorithm.}
\usage{
mre_fwd(target_set, evidence_set, evidence_states, bn_grain)
}
\arguments{
\item{target_set}{Set of variables to explore to explain the observed evidence}

\item{evidence_set}{Set of observed variable names}

\item{evidence_states}{Set of observed variable states}

\item{bn_grain}{Bayesian network as grain object}
}
\value{
Set of most relevant explanations
}
\description{
Solve the most relevant explanation in Bayesian networks using a forward search algorithm.
}
