% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mre_fwd_glasso.R
\name{mre_fwd_glasso}
\alias{mre_fwd_glasso}
\title{Solve the most relevant explanation in Bayesian networks using a forward-gLasso search algorithm.}
\usage{
mre_fwd_glasso(
  target_set,
  evidence_set,
  evidence_states,
  bn_grain,
  bn_rho,
  score_scale = TRUE
)
}
\arguments{
\item{target_set}{Set of variables to explore to explain the observed evidence}

\item{evidence_set}{Set of observed variable names}

\item{evidence_states}{Set of observed variable states}

\item{bn_grain}{Bayesian network as grain object}

\item{bn_rho}{L1 regularisation parameter for gLasso}

\item{score_scale}{Parameter to add very small random noise to score matrix for gLasso}
}
\value{
Set of most relevant explanations
}
\description{
Solve the most relevant explanation in Bayesian networks using a forward-gLasso search algorithm.
}
