% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbours.R
\name{neighbours}
\alias{neighbours}
\title{Title}
\usage{
neighbours(
  a1_neighbour_preglasso,
  c1_neighbour_preglasso,
  target_set,
  target_set_length,
  glasso_result_inv
)
}
\arguments{
\item{a1_neighbour_preglasso}{Add one neighbours}

\item{c1_neighbour_preglasso}{Change one neighbours}

\item{target_set}{Set of variables to explore to explain the observed evidence}

\item{target_set_length}{Number of variables in the target set}

\item{glasso_result_inv}{Result of gLasso}
}
\value{
List with two elements: all the neighbours, remaining neighbours after gLasso
}
\description{
Title
}
\keyword{internal}
