% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_matrix_fun.R
\name{score_matrix_fun}
\alias{score_matrix_fun}
\title{Create score matrix for gLasso}
\usage{
score_matrix_fun(df_target_neighbours, my_gbfs)
}
\arguments{
\item{df_target_neighbours}{Set of neighbours}

\item{my_gbfs}{Generalised Bayes factor scores}
}
\value{
Score matrix for gLasso
}
\description{
Create score matrix for gLasso
}
\keyword{internal}
