% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_names.R
\name{split_names}
\alias{split_names}
\title{Split names into nodes and states}
\usage{
split_names(names)
}
\arguments{
\item{names}{}
}
\value{
A list with node names and states
}
\description{
Split names into nodes and states
}
\keyword{internal}
